backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0

# Queue/Cache
redis>=5.2.0
rq>=2.0.0

# AI/LLM
anthropic>=0.39.0

# Google/Gmail API
google-api-python-client>=2.154.0
google-auth>=2.36.0
google-auth-oauthlib>=1.2.1

# Git Operations
gitpython>=3.1.43

# Browser Automation
playwright>=1.49.0

# PDF Parsing
pdfplumber>=0.11.0

# Auth
python-multipart>=0.0.18
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP & Utils
httpx>=0.28.0
aiofiles>=24.1.0
python-dotenv>=1.0.0

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
ruff>=0.8.0
mypy>=1.13.0
